term-project/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
massive
tqdm>=4.65.0

